+++ src/onegov.event/onegov/event/tests/test_cli.py
@@ -1,0 +1,113 @@
+from click.testing import CliRunner
+from onegov.core.utils import module_path
+from onegov.event.cli import cli
+from os import path
+from pytest import mark
+from unittest.mock import MagicMock
+from unittest.mock import patch
+from yaml import dump
+
+
+def write_cfg(postgres_dsn, session_manager, temporary_directory, redis_url):
+    cfg = {
+        'applications': [
+            {
+                'path': '/foo/*',
+                'application': 'onegov.core.Framework',
+                'namespace': 'foo',
+                'configuration': {
+                    'dsn': postgres_dsn,
+                    'redis_url': redis_url
+                }
+            }
+        ]
+    }
+
+    session_manager.ensure_schema_exists('foo-bar')
+
+    cfg_path = path.join(temporary_directory, 'onegov.yml')
+    with open(cfg_path, 'w') as f:
+        f.write(dump(cfg))
+
+    return cfg_path
+
+
+# test clear
+# test import ical
+# test import guidle
+
+
+@mark.parametrize("xml", [
+    module_path('onegov.event', 'tests/fixtures/guidle.xml'),
+])
+def test_import_guidle(
+    postgres_dsn, session_manager, temporary_directory, redis_url, xml
+):
+    cfg_path = write_cfg(
+        postgres_dsn, session_manager, temporary_directory, redis_url
+    )
+
+    runner = CliRunner()
+
+    with open(xml) as f:
+        text = f.read()
+    response = MagicMock(text=text)
+
+    # First import
+    with patch('onegov.event.cli.get', return_value=response):
+        result = runner.invoke(cli, [
+            '--config', cfg_path,
+            '--select', '/foo/bar',
+            'import-guidle', "'https://example.org/abcd'",
+        ])
+    assert result.exit_code == 0
+    assert "Events successfully imported" in result.output
+    assert "(4 added, 0 updated, 0 deleted)" in result.output
+
+    # Reimport
+    with patch('onegov.event.cli.get', return_value=response):
+        result = runner.invoke(cli, [
+            '--config', cfg_path,
+            '--select', '/foo/bar',
+            'import-guidle', "'https://example.org/abcd'",
+        ])
+    assert result.exit_code == 0
+    assert "Events successfully imported" in result.output
+    assert "(0 added, 0 updated, 0 deleted)" in result.output
+
+    # Clear
+    result = runner.invoke(cli, [
+        '--config', cfg_path,
+        '--select', '/foo/bar',
+        'clear'
+    ], 'y')
+    assert result.exit_code == 0
+
+    # Import
+    with patch('onegov.event.cli.get', return_value=response):
+        result = runner.invoke(cli, [
+            '--config', cfg_path,
+            '--select', '/foo/bar',
+            'import-guidle', "'https://example.org/abcd'",
+        ])
+    assert result.exit_code == 0
+    assert "Events successfully imported" in result.output
+    assert "(4 added, 0 updated, 0 deleted)" in result.output
+
+    # Create tagmap
+    tagmap = path.join(temporary_directory, 'tagmap.csv')
+    with open(tagmap, 'w') as f:
+        f.write("Konzert Pop / Rock / Jazz,Konzert\nSport,Sport")
+
+    # Re-import with tagmap
+    with patch('onegov.event.cli.get', return_value=response):
+        result = runner.invoke(cli, [
+            '--config', cfg_path,
+            '--select', '/foo/bar',
+            'import-guidle', "'https://example.org/abcd'",
+            '--tagmap', tagmap
+        ])
+    assert result.exit_code == 0
+    assert "Events successfully imported" in result.output
+    assert "Tags not in tagmap: \"Kulinarik\"!"
+    assert "(0 added, 4 updated, 0 deleted)" in result.output