+++ src/onegov.event/onegov/event/utils/__init__.py
@@ -1,0 +1,6 @@
+from onegov.event.utils.guidle import GuidleExportData
+
+
+__all__ = (
+    'GuidleExportData',
+)